Simulate slow service
This commit simulate a slow backend service and the impact on the Vaadin
UI. Because the RestTemplate call happens in the UI thread when it
initializes, no UI element is available until the remote actually
completes.

For the purpose of this demo, we're also showcasing `java.util.Duration`
support in Spring Boot that supports a friendly format (i.e. '5s' for
5 seconds) and metadata support.

Try typing demo.vaadin in `application.properties` from your favourite
IDE. You'll see the default value has been detected with the proper
format.

If you try to input an invalid value, the application won't start and
you'll get a failure analysis that tells you what went wrong, in which
file and where in the file. This is due to the new Origin support for
properties introduced in Spring Boot 2.