Add progress bar
To mitigate with this issue, let's add a progress bar that tells the
user what is going on. Notification easily allows you to "open" and
"close" a notification. In this commit we've add an indeterminate
ProgressBar to the notification.

Of course, the call is still part of the UI initialization so that
does not change anything. We should rather process that call
asynchronously and RestTemplate can be used for this if you take care
of using it in a separate thread.

Hopefully Spring 5 brings a nice alternative for this based on Reactive
streams.