Add WebClient alternative
This commit shows how the same feature can be implemented using the new
WebClient API in Spring Framework 5.

WebClient is a reactive, non-blocking client for HTTP requests with a
functional-style API client and Reactive Streams support. As you can see
from the method signature, we no longer return a list of Commit but a
`Flux` of Commit. `Flux` is an implementation of Publisher and allows
declarative composition of asynchronous logic.

This commit also showcases how such API can be tested.