Use WebClient
This commit showcases how to use the WebClient in our view. First we
make sure to collect all elements to a list (as a Mono of list). Then,
we use the `delayElement` operator to apply our fake delay. We want to
make sure the popup closes no matter what so we wrap that in the
`doAfterTerminate` operator. Note that as our call is happening outside
of the UI thread, we need to execute UI related code in a command that
we pass to the UI.

Because we are composing asynchronous logic, nothing happens until
something subscribe to the publisher. The last step is to subscribe to
the publisher and declare what should happen when the data is available
or when an error occurs respectively.