Automatically cancel pending request if another one is initiated
If the user calls refresh while a current HTTP request is being
processed, it would be nice to make sure the original request is
cancelled and swapped with the new one.

Fortunately, project reactor has a utility for that called `Swap`. When
you subscribe to a `Publisher` you get a `Subscription` instance that
allows you to manage it. In particular, you can cancel the subscription,
indicating the publisher you are no longer interested by the data.

This commit uses `Swap`. Whenever a new subscription is provided, swap
will automatically cancel any existing subscription. To showcase this,
this commit also adds the `log` operator on the chain that outputs what
the pipeline is doing.

Try starting the app and click refresh quickly. You'll see a "cancel"
event and a new request initiated.